Create A 101 : Structure of a C Program
What does a C Code look like?

C is a powerful language that can be used to achieve a lot of different tasks.

The code generally begins with #include statements. These statements include files that are built-in to C libraries and contains code for important tasks like taking input/output.

Execution of every program begins at the main() function, so in the body of the main function we write the code according to our requirements.

For your first task, write a code that prints “Hello World!” (without the double quotes).

Note that in C Language every logical statement ends with a semicolon ';'.

NOTE : You are expected to fix any other errors that you come across.