Create A 105 : Literals
in C we have literals or contant values along with variables. The values are nothing but the the direct values that we assign to any variable for eg. 5, 10, 2.5 are all constant values or literals.

Notice the "%d" in the previous code's printf statement? Use a similar printf statement here to print the value 10 to the console, but instead of storing 10 in a variable, print it directly.

NOTE : You are expected to fix any other errors that you come across.

Sample Output 0

10