Create A 118 : Shorthand Assignment Operator
To increase the value of a variable var by 1 or 2, we used var = var + 1 or var = var + 2.

C has a shortcut syntax for carrying out these tasks and they are known as shorthand operators. Their use is : var += 1 or var += 2. Infact any expression var = var + x can be shrinked to var += x;

This works also with other operators :
- var = var - x; -> var -= x;
- var = var * x; -> var *= x;
- var = var / x; -> var /= x;
- var = var % x; ->var %= x

To complete this task, input two numbers A and B, and print the first three terms of the series A + B, A + 2B, A + 3B;

Sample Input 0

5 3
Sample Output 0

8 11 14