Create Z 303 COUNT THE PRIME NUMBERS
Jayraj attended the campus placement drive conducted by Amazon. In the technical round the interviewer asked the candidate to write an algorithm to count the number of primes in a given range. Jayaraj wrote an algorithm as he is good at logic building. But the interviewer asked him to code it. But Jayaraj was not good at coding. So please help him in writing the code by submitting your solution to this problem

Input Format

Two space seperated Integers indicating the first and last values

Constraints

1<=first<=last<=10^12

Output Format

The count of prime number in that range

Sample Input 0

1 10
Sample Output 0

4
Explanation 0

The primes from 1 to 10 are 2,3,5,7