Create F M04 - Matrix Product
Write a program that takes as input two matrices and prints their product. If it is not possible to multiply the two matrices simply print -1.

Each matrix as input is of the form :

followed by number of lines with no. of elements on each line

More info on matrix multiplication : https://en.wikipedia.org/wiki/Matrix_multiplication

Hint : you will need three nested for loops to achieve the goal

Sample Input 0

2 2
1 0
0 1
2 2
1 1
1 1
Sample Output 0

1 1
1 1